Move remainder of fixtures functionality entirely within a single file (apply_fixtures.py). This should make it easier to refactor the logic.
