+++ easyvvuq/encoders/__init__.py
@@ -2,7 +2,7 @@ from .base import BaseEncoder
 from .generic_template import GenericEncoder
 from .multiencoder import MultiEncoder
 from .directory_builder import DirectoryBuilder
-from .fixtures import ApplyFixtures
+from .apply_fixtures import ApplyFixtures
 
 __copyright__ = """
